pyproject.toml
[project]
name = "mlops-multiagent"
version = "0.1.0"
description = "Sequential multi-agent flow for CV matching with LangGraph and LangChain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "pydantic>=2.7.0",
  "pypdf>=5.0.0",
  "python-docx>=1.1.0",
]

[project.scripts]
mlops-multiagent = "mlops_multiagent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mlops_multiagent"]